# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: HideAndSeek
  description: Geographic Hide and Seek game server
  version: 0.1.0
paths:
...
  /games/{game_id}:
    get:
      tags:
      - games
      summary: Get Game State
      description: Fetch current game state.
      operationId: get_game_state_games__game_id__get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /games/{game_id}/questions/radar:
    post:
      tags:
      - questions
      summary: Ask Radar Question
      description: Ask a radar question, spending a radar inventory slot.
      operationId: ask_radar_question_games__game_id__questions_radar_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskQuestionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/thermometer:
    post:
      tags:
      - questions
      summary: Ask Thermometer Question
      description: Ask a thermometer question, spending a thermometer inventory slot.
      operationId: ask_thermometer_question_games__game_id__questions_thermometer_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskQuestionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/matching:
    post:
      tags:
      - questions
      summary: Ask Matching Question
      description: Ask a matching question about a feature category.
      operationId: ask_matching_question_games__game_id__questions_matching_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskQuestionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/measuring:
    post:
      tags:
      - questions
      summary: Ask Measuring Question
      description: Ask a measuring question about a feature category.
      operationId: ask_measuring_question_games__game_id__questions_measuring_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskQuestionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/thermometer/{question_id}/lock-in:
    post:
      tags:
      - questions
      summary: Lock In Question
      description: Lock in the seeker's end position for a thermometer question.
      operationId: lock_in_question_games__game_id__questions_thermometer__question_id__lock_in_post
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Question Id
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/{question_id}/answer:
    post:
      tags:
      - questions
      summary: Answer Question
      description: "Hider answers a question \u2014 snapshot location, compute answer\
        \ and exclusion."
      operationId: answer_question_games__game_id__questions__question_id__answer_post
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Question Id
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions:
    get:
      tags:
      - questions
      summary: List Game Questions
      description: "Chronological list of all questions \u2014 whitelist summary only."
      operationId: list_game_questions_games__game_id__questions_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionSummaryResponse'
                title: Response List Game Questions Games  Game Id  Questions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/questions/{question_id}:
    get:
      tags:
      - questions
      summary: Get Question Detail
      description: "Full question detail \u2014 hider only."
      operationId: get_question_detail_games__game_id__questions__question_id__get
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Question Id
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/exclusions:
    get:
      tags:
      - questions
      summary: Get Exclusions
      description: "Seeker tactical map \u2014 per-question exclusion geometry."
      operationId: get_exclusions_games__game_id__exclusions_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExclusionsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AskQuestionRequest:
      properties:
        slot_index:
          type: integer
          title: Slot Index
          description: 0-based index into the inventory for this question type.
        location:
          $ref: '#/components/schemas/Point'
          description: Current seeker position as a GeoJSON Point.
        custom_distance:
          anyOf:
          - type: number
          - type: 'null'
          title: Custom Distance
          description: Required for custom slots (distance=null).
      type: object
      required:
      - slot_index
      - location
      title: AskQuestionRequest
      description: Ask a question, identified by slot_index. The URL determines question_type.
    AskQuestionResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        game_id:
          type: string
          format: uuid
          title: Game Id
        sequence:
          type: integer
          title: Sequence
          description: 1-based chronological order within the game.
        question_type:
          $ref: '#/components/schemas/QuestionType'
        status:
          $ref: '#/components/schemas/QuestionStatus'
        ask_count:
          type: integer
          title: Ask Count
          description: Which attempt this was (1 = first ask).
        parameters:
          anyOf:
          - $ref: '#/components/schemas/RadarParamsResponse'
          - $ref: '#/components/schemas/ThermometerParamsResponse'
          - $ref: '#/components/schemas/FeatureParamsResponse'
          title: Parameters
          description: Type-specific question parameters.
        asked_by:
          type: string
          format: uuid
          title: Asked By
          description: Player ID of the seeker who asked.
        asked_at:
          type: string
          format: date-time
          title: Asked At
        seeker_location_start:
          $ref: '#/components/schemas/Point'
          description: "GeoJSON Point \u2014 seeker position when asked."
      type: object
      required:
      - id
      - game_id
      - sequence
      - question_type
      - status
      - ask_count
      - parameters
      - asked_by
      - asked_at
      - seeker_location_start
      title: AskQuestionResponse
      description: "Slim response for ask endpoints \u2014 only fields meaningful\
        \ at ask time.\n\nNo answer-time fields (seeker_location_end, hider_location,\
        \ answered_at, answer)."
...
    ExclusionsResponse:
      properties:
        exclusions:
          items:
            $ref: '#/components/schemas/QuestionExclusionEntry'
          type: array
          title: Exclusions
        total_exclusion:
          anyOf:
          - oneOf:
            - $ref: '#/components/schemas/Point'
            - $ref: '#/components/schemas/MultiPoint'
            - $ref: '#/components/schemas/LineString'
            - $ref: '#/components/schemas/MultiLineString'
            - $ref: '#/components/schemas/Polygon'
            - $ref: '#/components/schemas/MultiPolygon'
            - $ref: '#/components/schemas/GeometryCollection'
            discriminator:
              propertyName: type
              mapping:
                GeometryCollection: '#/components/schemas/GeometryCollection'
                LineString: '#/components/schemas/LineString'
                MultiLineString: '#/components/schemas/MultiLineString'
                MultiPoint: '#/components/schemas/MultiPoint'
                MultiPolygon: '#/components/schemas/MultiPolygon'
                Point: '#/components/schemas/Point'
                Polygon: '#/components/schemas/Polygon'
          - type: 'null'
          title: Total Exclusion
          description: Cumulative exclusion across all answered questions.
      type: object
      required:
      - exclusions
      - total_exclusion
      title: ExclusionsResponse
      description: "Seeker tactical view \u2014 per-question exclusion geometry and\
        \ cumulative total."
...
    GameResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        map_id:
          type: string
          format: uuid
          title: Map Id
        status:
          $ref: '#/components/schemas/GameStatus'
        convention:
          type: string
          title: Convention
          description: 'Distance convention: "metric" or "imperial".'
        join_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Join Code
          description: 4-character code for joining. Null once hiding starts.
        hiding_time_min:
          type: integer
          title: Hiding Time Min
          description: Hiding phase duration in minutes.
        base_question_delay_min:
          type: integer
          title: Base Question Delay Min
          description: Auto-answer delay in minutes.
        inventory:
          $ref: '#/components/schemas/InventoryResponse'
          description: "Question inventory \u2014 all slots grouped by type."
        players:
          items:
            $ref: '#/components/schemas/PlayerResponse'
          type: array
          title: Players
        created_at:
          type: string
          format: date-time
          title: Created At
        hiding_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Hiding Started At
          description: When the hiding phase began. Clients compute countdown.
        seeking_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Seeking Started At
          description: When the seeking phase began.
      type: object
      required:
      - id
      - map_id
      - status
      - convention
      - join_code
      - hiding_time_min
      - base_question_delay_min
      - inventory
      - players
      - created_at
      title: GameResponse
      description: Full game state, including players and question inventory.
    GameStatus:
      type: string
      enum:
      - lobby
      - hiding
      - seeking
      - finished
      title: GameStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlayerResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        color:
          type: string
          title: Color
          description: Hex color, e.g. "#FF5733".
        role:
          anyOf:
          - $ref: '#/components/schemas/PlayerRole'
          - type: 'null'
          description: Null until the host assigns a role.
      type: object
      required:
      - id
      - name
      - color
      - role
      title: PlayerResponse
      description: A player in a game.
...
    QuestionDetailResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        game_id:
          type: string
          format: uuid
          title: Game Id
        sequence:
          type: integer
          title: Sequence
          description: 1-based chronological order within the game.
        question_type:
          $ref: '#/components/schemas/QuestionType'
        status:
          $ref: '#/components/schemas/QuestionStatus'
        ask_count:
          type: integer
          title: Ask Count
          description: Which attempt this was (1 = first ask).
        parameters:
          anyOf:
          - $ref: '#/components/schemas/RadarParamsResponse'
          - $ref: '#/components/schemas/ThermometerParamsResponse'
          - $ref: '#/components/schemas/FeatureParamsResponse'
          title: Parameters
          description: Type-specific question parameters.
        asked_by:
          type: string
          format: uuid
          title: Asked By
          description: Player ID of the seeker who asked.
        asked_at:
          type: string
          format: date-time
          title: Asked At
        seeker_location_start:
          $ref: '#/components/schemas/Point'
          description: "GeoJSON Point \u2014 seeker position when asked."
        seeker_location_end:
          anyOf:
          - $ref: '#/components/schemas/Point'
          - type: 'null'
          description: "GeoJSON Point \u2014 seeker position at lock-in (thermometer\
            \ only)."
        answered_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Answered At
        hider_location:
          anyOf:
          - $ref: '#/components/schemas/Point'
          - type: 'null'
          description: "GeoJSON Point \u2014 hider position at answer time."
        answer:
          anyOf:
          - type: string
          - type: 'null'
          title: Answer
          description: yes/no for radar, closer/farther for thermometer.
      type: object
      required:
      - id
      - game_id
      - sequence
      - question_type
      - status
      - ask_count
      - parameters
      - asked_by
      - asked_at
      - seeker_location_start
      - seeker_location_end
      - answered_at
      - hider_location
      - answer
      title: QuestionDetailResponse
      description: "Full question detail for hiders \u2014 everything except exclusion\
        \ geometry.\n\nUsed by the hider detail endpoint and write endpoints (answer/lock-in)."
...
    QuestionSummaryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        sequence:
          type: integer
          title: Sequence
          description: 1-based chronological order within the game.
        question_type:
          $ref: '#/components/schemas/QuestionType'
        status:
          $ref: '#/components/schemas/QuestionStatus'
        ask_count:
          type: integer
          title: Ask Count
          description: Which attempt this was (1 = first ask).
        asked_by:
          type: string
          format: uuid
          title: Asked By
          description: Player ID of the seeker who asked.
        asked_at:
          type: string
          format: date-time
          title: Asked At
        answered_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Answered At
        answer:
          anyOf:
          - type: string
          - type: 'null'
          title: Answer
          description: yes/no for radar, closer/farther for thermometer, etc.
      type: object
      required:
      - id
      - sequence
      - question_type
      - status
      - ask_count
      - asked_by
      - asked_at
      - answered_at
      - answer
      title: QuestionSummaryResponse
      description: "Lightweight question summary \u2014 whitelist of safe fields for\
        \ shared polling.\n\nNo parameters, no locations, no geometry. Both roles\
        \ use this to detect\nnew activity. New fields added to the question model\
        \ do not appear here\nuntil consciously included."
    QuestionType:
      type: string
      enum:
      - radar
      - thermometer
      - matching
      - measuring
      title: QuestionType
...